Fixed Bugs
- Added CAPS LOCK support.
- You can now press Enter to quickly go through the Join Game or Host
Game GUI.
- Fixed glitch with synchronization.
- Added Build mode tooltips.
- Asset cloning now has 2 levels, just a copy of the asset's internal
variables with the original base, or a copy of the asset's internal
variables as well as the base. The second way is more resource
intensive, but allows for on-the-fly editing of things like Textures,
while the first now allows for less resource intensive ways of cloning
basic assets.